requirements.txt
flask==3.1.0
supabase==2.13.0
python-dotenv==1.0.1
bcrypt==4.2.1
cryptography==44.0.0
google-genai>=1.0.0
sendgrid==6.11.0
twilio==9.4.0
requests==2.32.3
gunicorn==23.0.0
